+++ CHANGELOG.md
@@ -18,6 +18,7 @@ CentOS-6 6.9 x86_64, Apache 2.4, PHP-FPM 5.6, PHP memcached 2.2, Zend Opcache 7.
 - Adds test case output with improved readability.
 - Adds a healthcheck.
 - Removes healthcheck from functional tests of access log to prevent intermittent failures.
+- Fixes issue with local readonly variables being writable.
 
 ### 2.1.1 - 2017-03-12